Overrides getMergeInstance for Lucene99HnswVectorsReader and Lucene99FlatVectorsReader
Lucene99FlatVectorsReader is opened with IOContext READ with random access. This optimizes searches
with madvise as RANDOM. For merges we need sequential access and ability to preload pages to be able
to shorten the merge time. IOContext.READONCE gives sequential madvise to the kernel, which will preload the pages
aggressively and discard them once accessed.